hint message improved; logging dir renamed to logs_scorep_jupyter
